✅🤖 move mdim meta-tag e2e tests into their own file
Restores grapher-config-r2.e2e.test.ts to its original scope (the
.config.json R2 endpoint) and adds a test:functions:e2e-mdim script,
mirroring the existing per-file e2e scripts.
